+++ Core/test/Core/PipelineBuilderTests.cs
@@ -1,7 +1,7 @@
 // <copyright file="PipelineBuilderTests.cs" company="Chris Trout">
 // MIT License
 //
-// Copyright(c) 2019-2021 Chris Trout
+// Copyright © 2019-2021 Chris Trout
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -24,11 +24,11 @@
 
 namespace MyTrout.Pipelines.Core.Tests
 {
-    using Microsoft.Extensions.Logging;
+    using Microsoft.Extensions.Configuration;
+    using Microsoft.Extensions.DependencyInjection;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Moq;
     using MyTrout.Pipelines.Samples.Tests;
-    using MyTrout.Pipelines.Steps;
     using System;
     using System.Collections.Generic;
     using System.Globalization;
@@ -60,65 +60,24 @@ namespace MyTrout.Pipelines.Core.Tests
         }
 
         [TestMethod]
-        public void Does_Not_Throw_NullReferenceException_From_AddStep_When_StepContext_And_StepType_Are_Supplied_And_There_Are_No_StepAdded_Event_Listeners()
-        {
-            // arrange
-            var stepContext = "context-56";
-            var expectedStepType = typeof(SampleStep2);
-
-            var builder = new PipelineBuilder();
-
-            try
-            {
-                // act
-                builder.AddStep(expectedStepType, stepContext);
-            }
-            catch (Exception exc)
-            {
-                // assert
-                Assert.Fail(exc.ToString());
-            }
-        }
-
-        [TestMethod]
         public void Fires_StepAdded_Event_From_AddStep_When_GenericTypeParam_Is_Supplied()
         {
             // arrange
             var expectedType = typeof(SampleStep1);
 
             var builder = new PipelineBuilder();
-            builder.StepAdded += (object sender, StepAddedEventArgs e) =>
-                {
-                    // Note: This section is executed after the // act while still in-operation.
-                    // assert
+            builder.AddStepAddedEventHandler(
+                    (object sender, StepAddedEventArgs e) =>
+                    {
+                        // Note: This section is executed after the // act while still in-operation.
+                        // assert
                         Assert.AreEqual(builder, sender);
                         Assert.AreEqual(e.CurrentStep.StepType, expectedType);
                         Assert.IsNull(e.CurrentStep.StepContext);
-                };
+                    });
 
             // act
             builder.AddStep<SampleStep1>();
-        }
-
-        [TestMethod]
-        public void Fires_StepAdded_Event_From_AddStep_When_StepType_And_StepContext_Are_Supplied()
-        {
-            // arrange
-            var expectedStepContext = "context-482983";
-            var expectedStepType = typeof(SampleStep2);
-
-            var builder = new PipelineBuilder();
-            builder.StepAdded += (object sender, StepAddedEventArgs e) =>
-            {
-                // Note: This section is executed after the // act while still in-operation.
-                // assert
-                Assert.AreEqual(builder, sender);
-                Assert.AreEqual(e.CurrentStep.StepType, expectedStepType);
-                Assert.AreEqual(e.CurrentStep.StepContext, expectedStepContext);
-            };
-
-            // act
-            builder.AddStep(expectedStepType, expectedStepContext);
         }
 
         [TestMethod]
@@ -128,50 +87,29 @@ namespace MyTrout.Pipelines.Core.Tests
             var expectedType = typeof(SampleStep2);
 
             var builder = new PipelineBuilder();
-            builder.StepAdded += (object sender, StepAddedEventArgs e) =>
-            {
-                // Note: This section is executed after the // act while still in-operation.
-                // assert
-                Assert.AreEqual(builder, sender);
-                Assert.AreEqual(e.CurrentStep.StepType, expectedType);
-                Assert.IsNull(e.CurrentStep.StepContext);
-            };
+            builder.AddStepAddedEventHandler(
+                (object sender, StepAddedEventArgs e) =>
+                {
+                    // Note: This section is executed after the // act while still in-operation.
+                    // assert
+                    Assert.AreEqual(builder, sender);
+                    Assert.AreEqual(e.CurrentStep.StepType, expectedType);
+                    Assert.IsNull(e.CurrentStep.StepContext);
+                });
 
             // act
             builder.AddStep(expectedType);
         }
 
         [TestMethod]
-        public void Returns_Correct_PipelineRequest_Delegate_From_Build_Method()
-        {
-            // arrange
-            var nextRequest = new NoOpStep();
-            var logger = new Mock<ILogger<SampleWithInvokeAsyncMethodStep>>().Object;
-
-            var mockStepActivator = new Mock<IStepActivator>();
-            mockStepActivator.Setup(x => x.CreateInstance(It.IsAny<StepWithContext>(), It.IsAny<IPipelineRequest>()))
-                                            .Returns(new SampleWithInvokeAsyncMethodStep(logger, nextRequest));
-            IStepActivator stepActivator = mockStepActivator.Object;
-
-            var sut = new PipelineBuilder().AddStep<SampleWithInvokeAsyncMethodStep>();
-
-            var expectedTargetType = typeof(SampleWithInvokeAsyncMethodStep);
-
-            // act
-            var result = sut.Build(stepActivator);
-
-            // assert
-            Assert.IsNotNull(result);
-            Assert.IsInstanceOfType(result, expectedTargetType);
-        }
-
-        [TestMethod]
         public async Task Returns_Correct_Message_From_ConstructedPipeline()
         {
             // arrange
-            ILogger<StepActivator> logger = new Mock<ILogger<StepActivator>>().Object;
-            IServiceProvider serviceProvider = new Mock<IServiceProvider>().Object;
-            IStepActivator stepActivator = new StepActivator(logger, serviceProvider);
+            var mockStepActivator = new Mock<IStepActivator>();
+            mockStepActivator.Setup(x => x.CreateInstance(It.Is<StepWithContext>(x => x.StepType == typeof(SampleStep1)), It.IsAny<IPipelineRequest>())).Returns((StepWithContext step, IPipelineRequest next) => { return new SampleStep1(next); });
+            mockStepActivator.Setup(x => x.CreateInstance(It.Is<StepWithContext>(x => x.StepType == typeof(SampleStep2)), It.IsAny<IPipelineRequest>())).Returns((StepWithContext step, IPipelineRequest next) => { return new SampleStep2(next); });
+            mockStepActivator.Setup(x => x.CreateInstance(It.Is<StepWithContext>(x => x.StepType == typeof(SampleStep3)), It.IsAny<IPipelineRequest>())).Returns((StepWithContext step, IPipelineRequest next) => { return new SampleStep3(next); });
+            IStepActivator stepActivator = mockStepActivator.Object;
 
             var pipeline = new PipelineBuilder()
                                         .AddStep<SampleStep1>()
@@ -215,7 +153,7 @@ namespace MyTrout.Pipelines.Core.Tests
             // arrange
             var sut = new PipelineBuilder();
             Type stepType = typeof(int);
-            string expectedMessage = Resources.TYPE_MUST_BE_REFERENCE(CultureInfo.CurrentCulture, nameof(stepType));
+            string expectedMessage = Resources.TYPE_MUST_BE_REFERENCE(CultureInfo.CurrentCulture, "StepType");
 
             // act
             var result = Assert.ThrowsException<InvalidOperationException>(() => sut.AddStep(stepType));
@@ -231,7 +169,7 @@ namespace MyTrout.Pipelines.Core.Tests
             // arrange
             var sut = new PipelineBuilder();
             Type stepType = typeof(int);
-            string expectedMessage = Resources.TYPE_MUST_BE_REFERENCE(CultureInfo.CurrentCulture, nameof(stepType));
+            string expectedMessage = Resources.TYPE_MUST_BE_REFERENCE(CultureInfo.CurrentCulture, "StepType");
             string stepContext = "Step-Context-1";
 
             // act
@@ -248,7 +186,7 @@ namespace MyTrout.Pipelines.Core.Tests
             // arrange
             var sut = new PipelineBuilder();
             Type stepType = typeof(SampleWithoutIPipelineRequestStep);
-            string expectedMessage = Resources.TYPE_MUST_IMPLEMENT_IPIPELINEREQUEST(CultureInfo.CurrentCulture, nameof(stepType));
+            string expectedMessage = Resources.TYPE_MUST_IMPLEMENT_IPIPELINEREQUEST(CultureInfo.CurrentCulture, "StepType");
 
             // act
             var result = Assert.ThrowsException<InvalidOperationException>(() => sut.AddStep(stepType));
@@ -264,7 +202,7 @@ namespace MyTrout.Pipelines.Core.Tests
             // arrange
             var sut = new PipelineBuilder();
             Type stepType = typeof(SampleWithoutIPipelineRequestStep);
-            string expectedMessage = Resources.TYPE_MUST_IMPLEMENT_IPIPELINEREQUEST(CultureInfo.CurrentCulture, nameof(stepType));
+            string expectedMessage = Resources.TYPE_MUST_IMPLEMENT_IPIPELINEREQUEST(CultureInfo.CurrentCulture, "StepType");
             string stepContext = "Step-Context-1";
 
             // act
@@ -276,19 +214,31 @@ namespace MyTrout.Pipelines.Core.Tests
         }
 
         [TestMethod]
-        public void Throws_InvalidOperationException_From_AddStep_When_StepType_Is_Null()
+        public void Throws_InvalidOperationException_From_Build_When_A_StepType_Is_Null()
         {
+            // TO FUTURE DEVELOPER: THIS TEST REQUIRES KNOWLEDGE OF CLASS INTERNALS.
+            //                      IT IS INTENTIONALLY BRITTLE TO GUARANTEE THAT ANY
+            //                      VALIDATION ISSUES ARE CAPTURED.
+
             // arrange
-            var sut = new PipelineBuilder();
-            Type stepType = null;
-            string expectedParamName = nameof(stepType);
+            IStepActivator stepActivator = new Mock<IStepActivator>().Object;
+
+            var sut = new PipelineBuilder()
+                    .AddStep<SampleStep1>();
+
+            // Force a null into the PipelineBuilder's Step
+            PropertyInfo propertyInfo = typeof(PipelineBuilder).GetProperty("StepTypes", BindingFlags.NonPublic | BindingFlags.Instance);
+            var stepStack = propertyInfo.GetMethod.Invoke(sut, null) as Stack<StepWithContext>;
+            stepStack.Push(new StepWithContextUsingNullStep());
+
+            string expectedMessage = Resources.NULL_MIDDLEWARE(CultureInfo.CurrentCulture);
 
             // act
-            var result = Assert.ThrowsException<ArgumentNullException>(() => sut.AddStep(stepType));
+            var result = Assert.ThrowsException<InvalidOperationException>(() => sut.Build(stepActivator));
 
             // assert
             Assert.IsNotNull(result);
-            Assert.AreEqual(expectedParamName, result.ParamName);
+            Assert.AreEqual(expectedMessage, result.Message);
         }
 
         [TestMethod]
@@ -312,12 +262,13 @@ namespace MyTrout.Pipelines.Core.Tests
             Assert.AreEqual(expectedMessage, result.Message);
         }
 
-        // TO FUTURE DEVELOPER: THIS TEST REQUIRES KNOWLEDGE OF CLASS INTERNALS.
-        //                      IT IS INTENTIONALLY BRITTLE TO GUARANTEE THAT ANY
-        //                      VALIDATION ISSUES ARE CAPTURED.
         [TestMethod]
         public void Throws_InvalidOperationException_From_Build_When_StepTypes_Return_Null()
         {
+            // TO FUTURE DEVELOPER: THIS TEST REQUIRES KNOWLEDGE OF CLASS INTERNALS.
+            //                      IT IS INTENTIONALLY BRITTLE TO GUARANTEE THAT ANY
+            //                      VALIDATION ISSUES ARE CAPTURED.
+
             // arrange
             var mockStepActivator = new Mock<IStepActivator>();
             mockStepActivator.Setup(x => x.CreateInstance(It.IsAny<StepWithContext>(), It.IsAny<IPipelineRequest>())).Returns(new object());
@@ -341,33 +292,32 @@ namespace MyTrout.Pipelines.Core.Tests
             Assert.AreEqual(expectedMessage, result.Message);
         }
 
+        [TestCategory("Integration")]
         [TestMethod]
-        public async Task Validates_StepWithContext_Values_Passed_To_StepActivator_Using_Each_AddStep_OVerload()
+        public async Task Validates_StepWithContext_Values_Passed_To_StepActivator_Using_Each_AddStep_Overload()
         {
             // arrange
-            var logger = new Mock<ILogger<StepActivator>>().Object;
+            var config = new ConfigurationBuilder().AddJsonFile("test.json").Build();
 
-            Dictionary<string, SampleOptions> options = new Dictionary<string, SampleOptions>()
-            {
-                { "leading-context", new SampleOptions("leading-context") },
-                { "trailing-context", new SampleOptions("trailing-context") }
-            };
+            var serviceCollection = new ServiceCollection();
+            serviceCollection.AddScoped<IConfiguration>(_ => config);
+            serviceCollection.AddLogging();
+            serviceCollection.AddTransient<IStepActivator, StepActivator>();
 
-            var mockServiceProvider = new Mock<IServiceProvider>();
-            mockServiceProvider.Setup(x => x.GetService(typeof(IDictionary<string, SampleOptions>))).Returns(options);
-            var serviceProvider = mockServiceProvider.Object;
+            var serviceProvider = serviceCollection.BuildServiceProvider();
 
-            IStepActivator stepActivator = new StepActivator(logger, serviceProvider);
+            IStepActivator stepActivator = serviceProvider.GetRequiredService<IStepActivator>();
 
             var sut = new PipelineBuilder()
                             .AddStep<SampleStep1>()
                             .AddStep(typeof(SampleStep2))
                             .AddStep<SampleWithOptionsStep>("leading-context")
-                            .AddStep<SampleWithOptionsStep>("trailing-context")
+                            .AddStep(new StepWithContext(typeof(SampleWithOptionsStep), typeof(SampleOptions), "trailing-context"))
+                            .AddStep(new StepWithContext(typeof(SampleWithOptionsStep), typeof(SampleOptions), stepContext: "json-context", "last-context"))
                             .AddStep<SampleStep3>();
 
-            var expectedMessage = "Sponge Bob leading-context trailing-context SquarePants";
-            PipelineContext pipelineContext = new PipelineContext();
+            var expectedMessage = "Sponge Bob leading-connection-string trailing-connection-string last-connection-string SquarePants";
+            var pipelineContext = new PipelineContext();
 
             // act
             IPipelineRequest result = sut.Build(stepActivator);